complicated hedging function
hedging function in this commit uses a qubick function do determine our hedge_lot. the qubick function has been revised to use an inverse linear relationship if abs(etf_position) < 20, this was done to minimise our losses when we deem the market conditions as uncertain (ie. we're not committed to go 99 long/short). revisit this commit if we end up going down this (complex) heding function route after round 1.